fix: make DayPicker stateless, move themes to own file, remove unnecessary anonymous functions, change inner function to arrow function
